+++ pages/dashboard/index.tsx
@@ -65,7 +65,7 @@ const Dashboard = ({ cookie }) => {
                     </Box>
                 </Grid>
                 <Grid item xs={12} sx={{ margin: 'auto'}}>
-                    <AccountBalance />
+                    <AccountBalance cookie={cookie} />
                 </Grid>
             </Grid>
         </PageTemplateResponsive>